requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
async-timeout==4.0.2
attrs==23.2.0
base58==2.1.1
bitarray==2.6.0
blinker==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
ckzg==1.0.0
click==8.1.3
cytoolz==0.12.0
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.1.0
eth-utils==4.1.0
eth_abi==5.1.0
Flask==3.0.3
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
importlib-metadata==5.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
lru-dict==1.2.0
MarkupSafe==2.1.1
mnemonic==0.20
multiaddr==0.0.9
multidict==6.0.3
netaddr==0.8.0
parsimonious==0.10.0
protobuf==5.26.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
PyNaCl==1.5.0
pyrsistent==0.19.2
python-dotenv==1.0.1
pyunormalize==15.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-sse==0.3.1
rlp==4.0.0
rpds-py==0.18.0
six==1.16.0
stellar-sdk==9.3.0
toml==0.10.2
toolz==0.12.0
typing_extensions==4.11.0
urllib3==1.26.13
varint==1.0.2
waitress==3.0.0
web3==6.16.0
websockets==12.0
Werkzeug==3.0.2
xdrlib3==0.1.1
yarl==1.8.2
zipp==3.11.0
